optimize: Swarm memory - remove incomplete field, shrink complete/downloaded to u32 (72B→56B per swarm, -16B)

+++ src/types.rs
@@ -39,7 +39,7 @@ pub struct PeerState {
 #[derive(Clone, Debug, Default, Eq, PartialEq)]
 pub struct TorrentStats {
     pub complete: usize,
-    pub downloaded: u64,
+    pub downloaded: u32,
     pub incomplete: usize,
 }
